Added error title prefix and postfix support
Error description html documents can contain a <prefix> and a <postfix>
in their head section. The prefix will be displayed at the beginning
of an error message in the debugger state panel of the main GUI;
the postfix will be displayed at the end of the error message.

If there is no prefix defined, the document title will be taken.